No more "type error" exception
The test code for "terminal.py" now runs without crashing.  The output seems to get cut off after a certain string length is reached, which will not work for simulating a terminal environment: even just listing the files in the current directory results in a return string which gets cut off.  In the docs for the "Popen.communicate()" method, there is a note which says "The data read is buffered in memory, so do not use this method if the data size is large or unlimited" (https://docs.python.org/3/library/subprocess.html#subprocess.Popen).  So it seems that the issue is likely the use of the "Popen.communicate()" function when a more appropriate function should be used instead.